+++ kubejs/client_scripts/collection_tooltips.js
@@ -648,7 +648,8 @@ ItemEvents.tooltip(event => {
       'minecraft:dark_oak_sapling', 
       'minecraft:jungle_sapling', 
       'minecraft:oak_sapling', 
-      'minecraft:spruce_sapling'
+      'minecraft:spruce_sapling', 
+      'minecraft:mangrove_propagule'
     ],
     [
       'Flora Compendium',